add query-only column for SQL command 'SELECT length(BlobColumn), * FROM table'
SqConfig.h adds SQ_CONFIG_QUERY_ONLY_COLUMN to enable/disable this feature.
add macro SQ_N_ELEMENTS() to count number of elements present in an array.
macro SQ_N_PTRS() and SQ_N_COLUMNS() are deprecated.

add sample code of query-only column in storage-c99.c

+++ sqxc/SqType-buffer.c
@@ -48,7 +48,7 @@ int   sq_type_buffer_parse(void *instance, const SqType *type, Sqxc *src)
 		buf->writed = 0;
 		if (src->value.str == NULL)
 			break;
-		len = strlen(src->value.str);
+		len = (int)strlen(src->value.str);
 
 #if SQ_CONFIG_HAVE_JSONC
 		if (src->info == SQXC_INFO_JSONC_PARSER) {
